ln-simln-jamming/refactor: move batch_writer into own file
